+++ src/PortfolioHub.Domain/Exceptions/ValueObjects/InvalidQuantityException.cs
@@ -3,7 +3,7 @@
 public class InvalidQuantityException(string message) : BaseException(message)
 {
 
-    public static void ThrowIfInvalid(int value)
+    public static void ThrowIfInvalid(decimal value)
     {
         if (value < 0)
             ThrowIf(true, new InvalidQuantityException("A quantidade não pode ser negativa!"));
